+++ packages/geoview-core/public/templates/demos/demo-function-event.html
@@ -349,7 +349,7 @@
         <li>layer.onVisibleChanged for Water Quantity</li>
         <li>layer.onLayerOpacityChanged for Water Quantity</li>
         <li>layer.onLayerFilterApplied for Water Quantity</li>
-        <li>mapViewer.onGeolocatorSearch when a geolocator result is selected and zoomed to</li>
+        <li>mapViewer.controllers.mapController.onGeolocatorSearch when a geolocator result is selected and zoomed to</li>
       </ul>
     </div>
     <br />
@@ -415,7 +415,7 @@
         });
 
         // listen to geolocator search event
-        mapViewer.onGeolocatorSearch((sender, payload) => {
+        mapViewer.controllers.mapController.onGeolocatorSearch((sender, payload) => {
           mapViewer.notifications.addNotificationSuccess(
             `MapViewer: Geolocator search selected "${payload.searchItem}" at [${payload.coords[0].toFixed(4)}, ${payload.coords[1].toFixed(4)}]`
           );
